Add EveOnline::ESI::CharacterCalendarEvent
This adds support for getting the details of a given calendar event
from ESI.

Currently ESI has two endpoints which return event information
* /characters/{character_id}/calendar lists calendar event summaries
* /characters/{character_id}/calendar/{event_id} gets event details

We already use the `Event` model when listing calendar event summaries.
Thus despite it's name, the model represents an event's summary and not
an actual event.

A potential (re)use of the `Event` model in the context of getting event
details may be confusing because:
* the endpoints respond with different data structures
* the endpoints use different names to convey the same information. For
  example `event_response` vs. `response`

Because of the above, we add a new `EventDetails` model to represent
the details of an event.